CDDSO-641 Small refactor of WorkflowManager
